# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Habit Journey API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /memories:
    get:
      operationId: listMemories
      tags: [memories]
      summary: List memory journal entries
      parameters:
        - name: habitId
          in: query
          required: false
          schema: { type: integer }
      responses:
        "200":
          description: Memories
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Memory" }
    post:
      operationId: createMemory
      tags: [memories]
      summary: Create a memory (optionally with an uploaded photo)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MemoryInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Memory" }
  /memories/{memoryId}:
    delete:
      operationId: deleteMemory
      tags: [memories]
      summary: Delete a memory
      parameters:
        - name: memoryId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /rewards:
    get:
      operationId: listRewards
      tags: [rewards]
      summary: List rewards defined by the user
      responses:
        "200":
          description: Rewards
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Reward" }
    post:
      operationId: createReward
      tags: [rewards]
      summary: Create a reward
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RewardInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Reward" }
  /rewards/{rewardId}:
    patch:
      operationId: updateReward
      tags: [rewards]
      summary: Update a reward
      parameters:
        - name: rewardId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RewardUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Reward" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    delete:
      operationId: deleteReward
      tags: [rewards]
      summary: Delete a reward
      parameters:
        - name: rewardId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /rewards/{rewardId}/redeem:
    post:
      operationId: redeemReward
      tags: [rewards]
      summary: Redeem a reward by spending coins
      parameters:
        - name: rewardId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Redeemed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RedeemRewardResult" }
        "400":
          description: Insufficient coins or already redeemed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    Memory:
      type: object
      required: [id, habitId, note, photoUrl, date, createdAt]
      properties:
        id: { type: integer }
        habitId: { type: ["integer", "null"] }
        note: { type: string }
        photoUrl: { type: ["string", "null"] }
        date: { type: string, format: date }
        createdAt: { type: string, format: date-time }
    MemoryInput:
      type: object
      required: [note, date]
      properties:
        habitId: { type: integer }
        note: { type: string, minLength: 1 }
        photoObjectPath: { type: string, description: "objectPath returned by POST /storage/uploads/request-url" }
        date: { type: string, format: date }
    Reward:
      type: object
      required: [id, habitId, title, emoji, coinCost, isRedeemed, redeemedAt, createdAt]
      properties:
        id: { type: integer }
        habitId: { type: ["integer", "null"] }
        title: { type: string }
        emoji: { type: string }
        coinCost: { type: integer }
        isRedeemed: { type: boolean }
        redeemedAt: { type: ["string", "null"], format: date-time }
        createdAt: { type: string, format: date-time }
    RewardInput:
      type: object
      required: [title, emoji, coinCost]
      properties:
        habitId: { type: integer }
        title: { type: string, minLength: 1 }
        emoji: { type: string }
        coinCost: { type: integer, minimum: 1 }
    RewardUpdate:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        emoji: { type: string }
        coinCost: { type: integer, minimum: 1 }
...
    RedeemRewardResult:
      type: object
      required: [reward, coinsRemaining]
      properties:
        reward: { $ref: "#/components/schemas/Reward" }
        coinsRemaining: { type: integer }
...